+++ tests/fixer_test/.phan/config.php
@@ -135,5 +135,6 @@ return [
         'NotFullyQualifiedUsagePlugin',
         'PHPDocToRealTypesPlugin',
         'PHPDocRedundantPlugin',
+        'DeprecateAliasPlugin',
     ],
 ];