+++ tests/fixer_test/expected_src/008_deprecated_alias.php
@@ -1,0 +1,4 @@
+<?php
+return [
+ count([]),\implode(['a','b']),
+];