+++ tests/fixer_test/src/008_deprecated_alias.php
@@ -1,0 +1,4 @@
+<?php
+return [
+ sizeof([]),\join(['a','b']),
+];